+++ src/webpack/module.js
@@ -1,12 +1,12 @@
-import Utils from '../utils/';
-import getBabelConfig from './getBabelConfig';
+import Utils from '../utils/'
+import getRules from './getRules'
 export default function() {
     return {
-        rules: getBabelConfig(Utils.getUserConfig),
+        rules: getRules(Utils.getUserConfig),
         //不在扫正则所匹配的模块的依赖
         noParse: function(content) {
-            const pass = /(lodash|zepto|jquery)\.js/.test(content);
-            return pass;
+            const pass = /(lodash|zepto|jquery)\.js/.test(content)
+            return pass
         }
-    };
+    }
 }